requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2
python-dotenv==1.0.1
Pillow==10.4.0
drf-nested-routers==0.94.1